+++ create_addon_zip.ps1
@@ -11,7 +11,7 @@ Write-Host ""
 
 # Set variables
 $AddonName = "CrossRig"
-$Version = "1.1.0"
+$Version = "1.2.0"
 $ReleaseDir = "release"
 $TempDir = "${AddonName}_temp"